# swagger.yaml
openapi: 3.0.0
info:
  title: Growcoins Backend API
  version: 1.0.0
  description: |
    Comprehensive API documentation for Growcoins fintech application.
    This API provides endpoints for user management, goals, investments, savings, and more.
  contact:
    name: Growcoins API Support
    email: support@growcoins.com
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.growcoins.com
    description: Production server
...
paths:
...
  /api/risk-assessment:
    post:
      tags:
        - Risk Assessment
      summary: Save risk assessment
      description: Submit user's risk assessment answers and get risk profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - answers
                - total_score
                - risk_profile
              properties:
                user_id:
                  type: integer
                answers:
                  type: array
                  minItems: 5
                  maxItems: 5
                  items:
                    type: object
                    properties:
                      questionId:
                        type: integer
                      optionId:
                        type: string
                      answerText:
                        type: string
                      score:
                        type: integer
                        minimum: 1
                        maximum: 4
                total_score:
                  type: integer
                  minimum: 5
                  maximum: 20
                risk_profile:
                  type: string
                  enum: [Conservative, Moderate, Moderately Aggressive, Aggressive]
                recommendation:
                  type: string
      responses:
        '201':
          description: Risk assessment saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  assessment:
                    $ref: '#/components/schemas/RiskAssessment'
        '400':
          description: Validation error
        '404':
          description: User not found
...
  /api/goals:
    post:
      tags:
        - Goals
      summary: Create a new goal
      description: Create a new financial goal with automatic investment plan generation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - category_id
                - goal_name
                - target_amount
                - target_date
              properties:
                user_id:
                  type: integer
                category_id:
                  type: integer
                goal_name:
                  type: string
                  minLength: 3
                  maxLength: 100
                target_amount:
                  type: number
                  minimum: 100
                target_date:
                  type: string
                  format: date
                initial_amount:
                  type: number
                  minimum: 0
                frequency:
                  type: string
                  enum: [monthly, weekly, daily]
                monthly_amount:
                  type: number
                  minimum: 100
      responses:
        '201':
          description: Goal created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  goal:
                    $ref: '#/components/schemas/Goal'
        '400':
          description: Validation error
        '404':
          description: User not found
...
components:
  schemas:
...
    Goal:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        category_id:
          type: integer
        goal_name:
          type: string
        target_amount:
          type: number
        current_amount:
          type: number
        target_date:
          type: string
          format: date
        status:
          type: string
          enum: [active, completed, paused, cancelled]
        progress_percentage:
          type: number
        days_remaining:
          type: integer
        monthly_savings_needed:
          type: number
...
    RiskAssessment:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        total_score:
          type: integer
        risk_profile:
          type: string
          enum: [Conservative, Moderate, Moderately Aggressive, Aggressive]
        recommendation:
          type: string
        completed_at:
          type: string
          format: date-time
        answers:
          type: array
          items:
            type: object
            properties:
              question_id:
                type: integer
              option_id:
                type: string
              answer_text:
                type: string
              score:
                type: integer
...